requirements.txt
langchain-community>=0.4.1
langchain-chroma>=1.0.0
langchain-openai>=1.0.1
python-dotenv>=1.0.0
langchain>=1.0.3
